refactor(Dockerfile): remove commented librealsense installation
This commit removes the commented-out code related to the installation of
librealsense in the Dockerfile. The commented code included steps for adding
the Intel RealSense repository key, adding the repository, installing
librealsense packages, and verifying kernel updates for RealSense. These
steps are no longer necessary and have been removed to clean up the Dockerfile.